create the study plan service
