+++ apps/study_plan_service/src/config.py
@@ -1,4 +1,4 @@
 KAFKA_BROKER_URL = "localhost:9092"
 KAFKA_REQUEST_TOPIC = "study-plan-requests"
 KAFKA_RESULT_TOPIC = "study-plan-results"
-DATABASE_URL = "postgresql://user:password@host:port/dbname"
+DATABASE_URL = "postgresql://postgres:SEP-Group01@cse@db.cfcgwqdbnftjqsbgmsun.supabase.co:5432/postgres"
