+++ apps/study_plan_service/src/services/plan_generator.py
@@ -1,4 +1,4 @@
-from models.study_plan import StudyPlanRequest, StudyPlanResult
+from src.models.study_plan import StudyPlanRequest, StudyPlanResult
 
 def generate_study_plan(request: StudyPlanRequest, free_slots):
     # Simple logic: fill slots, calculate weeks required
